requirements.txt
about-time==2.0.4
aiohttp==3.4.4
aioredis==1.1.0
amqp==2.3.2
asgiref==3.2.3
astroid==2.0.4
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
billiard==3.5.0.4
bleach==3.0.1
celery==4.2.1
Cerberus==1.2
certifi==2018.8.24
channels==2.3.1
channels-redis==2.3.1
chardet==3.0.4
clearly==0.6.2
Click==7.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==2.2.2
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.2
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
entrypoints==0.2.3
flower==0.9.2
gevent==1.3.6
greenlet==0.4.15
grpcio==1.15.0
gunicorn==19.9.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
idna-ssl==1.1.0
incremental==17.5.0
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itypes==1.1.0
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kombu==4.2.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.4
msgpack==0.5.6
multidict==4.4.2
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
openapi-codec==1.3.2
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.4.1
prompt-toolkit==2.0.5
protobuf==3.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
PyHamcrest==1.9.0
pylint==2.1.1
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.1.2
pycrypto==2.6.1
paramiko==2.6.0
qtconsole==4.4.1
redis==2.10.6
requests==2.20.0
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
Twisted==18.7.0
txaio==18.8.1
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
websocket==0.2.1
websocket-client==0.53.0
websockets==6.0
widgetsnbextension==3.4.2
wrapt==1.10.11
yarl==1.2.6
zope.interface==4.5.0
jumpssh==1.5.1
bootstrap_admin==0.4.0
jsonfield==2.0.2
#logging==0.4.9.6
django-celery-beat==1.4.0
pyyaml== 5.1.0
http_parser
#django_json_widget
